+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - This Change Log!
 - `Sprinkles.QueryString.parse()` to parse query string params from any URL.
 - `location.$params` returns the params in `window.location.search`.
+- `Number.prototype.$ordinalize()` (moved from `String.prototype`).
 
 ### Changed
 - Cookie-related methods `setItem()`, `getItem()`, and `removeItem()` have been
@@ -22,6 +23,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `Location.$getParam()`
 - `Location.prototype.$getParams()`
 - `Location.prototype.$getParam()`
+- `String.prototype.$ordinalize()`
 
 ## [1.4.0] - 2014-07-09
 ### Added
@@ -45,7 +47,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `Location.$getParam()`
 - `Location.prototype.$getParams()`
 - `Location.prototype.$getParam()`
-- `String.prototype.$ordinalize()`
 - `window.$get()`
 - `window.$getJSON()`